Refactored Driver so that it is more static orientated. As theres no reason why not. There should only be one driver per running program.
Addition of ConcurrentSafeArrayList which is used in Driver. Basically allows for concurrent access to an arraylist.

Major update for the Level Editor. Is pretty much done, All we need now is to add the actual level parts

Oh and refactored the level folder so that levels are stored in assets/level but the individual images for each "level part" are in assets/levelparts